Attempted a fix for renaming however this failed,
renaming can only successfully take place on mobiles by touching another
part of the screen to remove the keyboard after typing the new name, if
the OK button is touched then the name reverts to the previous value.
The fix that was attempted was the removal of the OK button, however the
code required to do this has lost it's functionality in Unity 5, added
"TouchScreenKeyboard.hideInput = true;" (previously mentioned code used
to hide the OK button) to the cow OnGUI() function just in case this
issue gets fixed in a later version of Unity. Another thing that was
added was a line of code that stops random characters from being put
into cow names, only letters/numbers and spaces can be entered now.