Updated the new GUI, removed most of old GUI, the shop buttons still remain in order to purchase upgrades.
Added inventory to new GUI and added the delete buttons to the inventory
slots, still need to work on hiding the delete buttons for the slots
that are empty. Fixed the positioning of the new GUI so now it will
correctly position, it should also scale to larger resolutions but this
has not been tested yet, it does however scale to smaller resolutions as
this has been tested using the editor. Scaling to larger resolutions
should theoretically work seeing as the exact same textures have been
used in the new GUI that have scaled successfully in the old GUI.