Added a real time day/night rotation cycle for the sky, it is currently commented out however as it has not been
discussed with the rest of the team. The sky image file has been rotated
by 180 degrees, but the rotation value in unity has been set to 180 to
display it with the sun on the top, the reason for the rotation was for
the real time testing as the rotation in real time uses the system's
time, and an angle of 0 matched a time of 00:00, which is in the middle
of the night and not day (thus needing the moon at the top).